Improve import script

+++ src/import-sqlite.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 set -e
 
-rm transit.db
-rm  google_transit.zip
-rm -r feed
+[ -f transit.db ] && rm transit.db
+[ -f google_transit.zip ] && rm  google_transit.zip
+[ -d feed ] && rm -r feed
 curl -O http://www.transperth.wa.gov.au/TimetablePDFs/GoogleTransit/google_transit.zip
 mkdir feed
 cd feed
